+++ CHANGELOG.md
@@ -9,7 +9,7 @@
 - Use Bract 0.4.0
   - [TODO] Use the GA version of Bract 0.4.0
 - Bump CLJS version to 1.9.671
-- [Todo] Move stateful vars from `config` to `global` namespace
+- Move global/stateful vars from `config` to `global` namespace
 - [TODO] Rename namespace `demo.todomvc.config` to `demo.todomvc.keydef`
 - [TODO] Implement `version` CLI command